code cleanup...  constify/use references.
	modified:   include/alubox_fu_parts.h
	modified:   include/alubox_fu_s_parts.h
	modified:   include/alubox_switch_case.h
	modified:   include/bitfield.h
	modified:   include/bitfield_pb.h
	modified:   include/bitops.h
	modified:   include/bounds.h
	modified:   include/bswap.h
	modified:   include/callback_list.h
	modified:   include/callback_qlist.h
	modified:   include/capstone_assert_test.h
	modified:   include/debug.h
	modified:   include/dtime.h
	modified:   include/err_test.h
	modified:   include/handle.h
	modified:   include/ldst_stdint.h
	modified:   include/log.h
	modified:   include/mem_access.h
	modified:   include/mem_access_be.h
	modified:   include/mem_access_le.h
	modified:   include/qbasic.h
	modified:   include/queue.h
	modified:   include/shift_roll.h
	modified:   include/spinlock.h
	modified:   include/sub64.h
	modified:   include/trace.h
	modified:   include/typeof_access.h
	modified:   include/unused.h
	modified:   source/action.c
	modified:   source/callback_list.c
	modified:   source/callback_qlist.c
	modified:   source/dtime.c
	modified:   source/queue.c

+++ include/mem_access_le.h
@@ -15,8 +15,9 @@
 
 /* **** */
 
-static inline unsigned mem_0x0le__access_read(void *const p2src, const size_t size) {
-	__test_assert_size(sizeof(unsigned) <= size);
+static inline
+unsigned mem_0x0le__access_read(void *const p2src, const size_t size) {
+	__test_assert_size(sizeof(long) <= size);
 
 	uint32_t res = 0;
 	uint8_t* src = (uint8_t*)p2src;
@@ -27,7 +28,8 @@ static inline unsigned mem_0x0le__access_read(void *const p2src, const size_t si
 	return(res);
 }
 
-static inline void mem_0x0le__access_write(void *const p2dst, const size_t size, const uint32_t data) {
+static inline
+void mem_0x0le__access_write(void *const p2dst, const size_t size, const uint32_t data) {
 	__test_assert_size(sizeof(unsigned) <= size);
 
 	uint8_t* dst = (uint8_t*)p2dst;
@@ -36,8 +38,9 @@ static inline void mem_0x0le__access_write(void *const p2dst, const size_t size,
 		*dst++ = (data >> (i << 3)) & 0xff;
 }
 
-static inline unsigned mem_0x0le_access(void *const p2pat, const size_t size, const unsigned *const write) {
-	const unsigned data = write ? *write : mem_0x0le__access_read(p2pat, size);
+static inline
+unsigned mem_0x0le_access(void *const p2pat, const size_t size, const unsigned *const write) {
+	const long data = write ? *write : mem_0x0le__access_read(p2pat, size);
 
 	if(write)
 		mem_0x0le__access_write(p2pat, size, data);
@@ -47,7 +50,8 @@ static inline unsigned mem_0x0le_access(void *const p2pat, const size_t size, co
 
 /* **** */
 
-static inline unsigned mem_16le_access(void *const p2pat, const unsigned *const write) {
+static inline
+unsigned mem_16le_access(void *const p2pat, const unsigned *const write) {
 	const uint16_t data = (uint16_t)(write ? *write : htole16(*(uint16_t*)p2pat));
 
 	if(write)
@@ -56,8 +60,9 @@ static inline unsigned mem_16le_access(void *const p2pat, const unsigned *const 
 	return(data);
 }
 
-static unsigned mem_32le_access(void *const p2pat, const unsigned *const write) {
-	const unsigned data = write ? *write : htole32(*(uint32_t*)p2pat);
+static
+unsigned mem_32le_access(void *const p2pat, const unsigned *const write) {
+	const long data = write ? *write : htole32(*(uint32_t*)p2pat);
 
 	if(write)
 		*(uint32_t*)p2pat = le32toh(data);
@@ -65,7 +70,8 @@ static unsigned mem_32le_access(void *const p2pat, const unsigned *const write) 
 	return(data);
 }
 
-static inline unsigned mem_64le_access(void *const p2pat, const unsigned *const write) {
+static inline
+unsigned mem_64le_access(void *const p2pat, const unsigned *const write) {
 	if(sizeof(unsigned) <= sizeof(uint64_t))
 		return(mem_0x0le_access(p2pat, sizeof(uint64_t), write));
 
@@ -74,12 +80,13 @@ static inline unsigned mem_64le_access(void *const p2pat, const unsigned *const 
 	if(write)
 		*(uint64_t*)p2pat = le64toh(data);
 
-	return((unsigned)data);
+	return((long)data);
 }
 
 /* **** */
 
-static inline unsigned mem_access_le(void *const p2pat, const size_t size, const unsigned *const write) {
+static inline
+unsigned mem_access_le(void *const p2pat, const size_t size, const unsigned *const write) {
 	switch(size) {
 //		case 8:
 //			return(mem_64le_access(p2pat, write));
